Make population building much faster.
GOMAXPROCS is now set to the number of CPUs.

The RandomVariableOrdering heuristic was changed to select values from
the domain at random. Checking for the optimal element in each case is
very costly and this lead to a speedup from 1hr 20min runtime to 45s.

Added Population.Best() to retrieve the best solution in a population.