Version 0.5. Add fetch subcommand. Refactor heuristics.
The fetch subcommand replaces the get-instances.sh shell script. This
was done so that a shell is not required to automatically download the
ITC 2007 instances.

The MostConstrainedOrdering and RandomVariableOrdering were refactored
to use the new Solution.Best(e), which assigns to e the best element in
its domain. These functions are now just wrappers around loops of
Solution.Best.

The Population methods Population.RemoveOne and Population.Pick no
longer return the value of the solution -- only the solution itself.