Replace Solution.Clone() and change *Solution to []Rat in messages
Solution.QuickAssign was renamed to Solution.TryAssign; this name
reflects what it does better.

Rename Rat.assigned to Rat.Assigned so that it may be exported.

[]Rat is more lightweight than *Solution and doesn't require as much
processing to pass messages around. Solution.Assignments() makes a copy
of the solution's internal rats slice and Instance.SolutionFromRats
builds a solution out of a rats slice.

Solution.Clone was removed as it is no longer needed.