Remove dead code from Solution.makeDomain.
Solution.makeDomain contained extra code from when Domains were not
generated with the Solution itself (for partial domains etc). Now that
we have shrink / unshrink with assign and unassign, we can remove this
dead code.

Solution.makeDomain and solution.makeDomains have been refactored into
Instance.NewSolution so that all the code for creating a solution is in
Instance.NewSolution.