Version 0.7
Option parsing has been moved into a new package: spaghetti/options.

There are now --minpop and --maxpop flags that allow you to set the
minimum and maximum population for the HPGA populations. The defaults
are now 50 and 75.

The default number of islands and slaves are now 2 and 2.

Instance.ParseSolution has been refactored to use
Instance.SolutionFromRats since they shared most of the same
functionality.

Removed -x from .go files and from protocol.md

The controller now supports handling SIGINT. If SIGINT is received while
the controller is running, it will return the current best solution and
exit.

The value of the solution is printed at exit.
Now print the value of solution at exit.

In island.go, the sending of solution messages was incorrect; it sent
the []tt.Rat before the tt.Value, which was incorrect. This was not
caught until the program found a better than initial result.

Moved operation probabilities into solver/hpga/operations.go. The
probabilities now read more clearly in solver/hpga/slave.go.