Added naive GM operator.
The naive GM operator adds a timer to the island's message loop. Every 5
minutes, the island asks for a solution from a slave and performs GM on
it and sends it back to the child. The GM is performed by clearing all
conflicts, doing a mutate and assigning a best value to a variable, and
follwing it up with a forward search using the most-constrained variable
first heuristic.