The student attendance matrix now indicates event being attended
Instead of just indicating whether or not a student attends an event (via
a boolean), we now indicate the presence of an event by a non-negative
integer and the absence of an event by -1.

Solution.Fitness() conseuctive classes calculation simplified by one
branch.

Added Solution.AssignmentQuality() which determines how many soft
constraints an assignment violates. It is not the case that the sum of
assignment qualities is Solution.Value() because each event involved in
a conflict will be counted as many times as there are events in the
conflict.