Mutation is now the in the population package.
The population package has an exported function Mutate() which takes a
solution and mutates up to 20% of it randomly. The value of the solution
is returned.

The Population type now has an exported method MutateOne() which does
RemoveOne(), Mutate(), and insert().

newIndividual now takes optional value parameter. If the value parameter
is provided, then that value will be used as the value field in the
returned *individual. Otherwise, it will be calculated (as before).

Add Population.insert which takes a solution and a value.
Population.Insert calls Population.insert after calculating a Solution's
value. This was added so that a call to MutateOne() would not result in a
Solution's value being calculated twice.