Refactor Solution.AssignmentQuality
Solution.AssignmentQuality is new Solution.AssignmentQualities. The new
Solution.AssignmentQuality takes an event as a parameter and only computes
the quality for that event.

Solution.AssignmentQuality no longer iterates over the whole attendance
matrix for each event. It only iterates over the students that in fact
attend the event. This is faster and should give a more accurate estimate
of assignment quality.

Solution.AssignmentQuality now gives a more accurate estimate of
assignment quality by giving a result more similar to that of
Solution.Violations (i.e. it takes pairs of events into account and number
of events assigned to a room and time).