Add Solution.FindImprovements.
Solution.FindImprovements intends to find an event assignment that can be
improved in the given solution. It does so by calling Solution.improve on
each event. However, it iterates through the events in a random
permutation and so the first one that it finds is the one it returns. This
is equivalent to doing Solution.improve on each element and picking one
randomly.

Solution.Improve is renamed to Solution.improve. It is only called by
Solution.FindImprovements.