Version 0.9: Populations are island-wide.
Populations are no longer confined to individual slaves. Populations are
now a large array. Each slave gets a sub-population which is a slice of
that array. When a slave fills up its slice, it signals the island that it
is ready for selection (via a fullMessage) and waits for the island to
tell it to continue (via a continueMessage).