Version 0.11: Implement GM operator
The GM operator runs as a goroutine under each island. When an island
receives an orderingMessage from the controller, it signals that the GM
should start producing individuals. As it stands, the GM will produce 5%
of NSlaves * MinPop (rounded down). Selection has changed so that it takes
a slice of individuals to add into the populations. The selection
algorithm will pick less individuals for selection and insert individuals
from the GM so that the total population is still NSlaves * MinPop. After
selection, the GM is signalled to produce more individuals for the island.

Increase the values for maxTries and cutOff in hopes of generating better
individuals.