Individuals generated by GM are considered.
Previously, it was possible the GM would generate better individuals than
the best known individual and they would not be checked if they were. Now
when an island does selection, it loops through the list of generated
solutions and compares their values to the best-known values. If so, it
handles them by sending a valueMessage to its children and a
solutionMessage to the controller.