Rename Instance.domains to Instance.Domains.
Instance.Domains are now exported so that they can be used without
generating a new solution (for hill climbing).

Add the OrderedWeightedAssignment heuristic which assigns to a solution
based on weights determined by a meta-heuristic.

Add Solution.MakeShrinkableDomains which generates a list of domains (as
sets) that can be shrunk.